+++ Koha/Plugin/Com/ByWaterSolutions/RapidoILL/lib/RapidoILL/Backend/BorrowerActions.pm
@@ -278,7 +278,9 @@ sub item_in_transit {
 
                 $req->status('B_ITEM_IN_TRANSIT')->store;
 
-                $self->{plugin}->get_client( $self->{pod} )->borrower_item_returned( { circId => $circId }, $options );
+                if ( !exists $options->{notify_rapido} || $options->{notify_rapido} ) {
+                    $self->{plugin}->get_client( $self->{pod} )->borrower_item_returned( { circId => $circId }, $options );
+                }
             }
         );
         return $self;